refactor: Extract helper classes and streaming logic. #2808
